refactor: Refactor related class to slack package
