+++ src/main/java/tw/b2e/slack/SlackAppController.java
@@ -1,4 +1,4 @@
-package tw.b2e;
+package tw.b2e.slack;
 
 import com.slack.api.bolt.App;
 import com.slack.api.bolt.servlet.SlackAppServlet;
